Version 1.0
Since vim 7.0 supports :smap (mappings that apply only to Select mode) I updated the script to use these.  Now the maps defined by the script for when completion is active will not do anything unexpected in Visual mode.

I also updated the installation instructions in the comments at the top of the file.